Major 2020 Update Final
As far as I know, this will be the final functionality update I do. I don't think anything more has to be done.
Changelog:
-Fixed both previously known bugs. They shouldn't bother us anymore.
-Changed the placeholder title text. It would look better with a logo, but have fun trying to get an artist for us.
-Added plenty of documentation to the code through comments. It should be a lot easier to understand now.
Whoever gets to pick this up from he, have fun with this. I hope it doesn't cause you as much of a headache as it did for me.